feat: Implement the search page functionalities.
This commit implements the basic functionality of allowing
the user to browse books from the search page. The user
could also change the status of books from that page.
However, the search function is returning arbitrary result
for the purpose of testing and this will need to be implemented
properly in the next commits.